fix(tests): Move mock_global_access and mock_self_access to top level conftest.py
